+++ ingenialink/ethercat/network.py
@@ -213,7 +213,6 @@ class EthercatNetwork(Network):
         if slave_id not in self.__last_init_nodes:
             raise ILError(f"Slave {slave_id} was not found.")
         slave = self._ecat_master.slaves[slave_id - 1]
-        slave.add_emergency_callback(self._emcy_callback)
         servo = EthercatServo(
             slave, slave_id, dictionary, self._connection_timeout, servo_status_listener
         )
@@ -540,10 +539,3 @@ class EthercatNetwork(Network):
             )
         logger.warning(log_message)
         return all_drives_in_preop
-
-    def _emcy_callback(self, emergency_msg: "pysoem.Emergency") -> None:
-        """Log the emergency messages"""
-        slave_id = emergency_msg.slave_pos
-        if servo := next((servo for servo in self.servos if servo.slave_id == slave_id), None):
-            error_description = servo.get_emergency_description(emergency_msg.error_code)
-            logger.warning(f"Emergency message received from slave {slave_id}: {error_description}")